+++ lib/core/catcher_screenshot_manager.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui' as ui;
 
-import 'package:catcher/utils/catcher_logger.dart';
+import 'package:athmany_catcher/utils/catcher_logger.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';