+++ lib/handlers/base_email_handler.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
-import 'package:catcher/model/report.dart';
-import 'package:catcher/model/report_handler.dart';
+import 'package:athmany_catcher/model/report.dart';
+import 'package:athmany_catcher/model/report_handler.dart';
 
 ///Base class for all email handlers.
 abstract class BaseEmailHandler extends ReportHandler {
@@ -45,9 +45,7 @@ abstract class BaseEmailHandler extends ReportHandler {
     if (enableStackTrace) {
       buffer.write("<h2>Stack trace:</h2>");
 
-      _escapeHtmlValue(report.stackTrace.toString())
-          .split("\n")
-          .forEach((element) {
+      _escapeHtmlValue(report.stackTrace.toString()).split("\n").forEach((element) {
         buffer.write("$element<br>");
       });
       buffer.write("<hr><br>");
@@ -55,16 +53,14 @@ abstract class BaseEmailHandler extends ReportHandler {
     if (enableDeviceParameters) {
       buffer.write("<h2>Device parameters:</h2>");
       for (final entry in report.deviceParameters.entries) {
-        buffer
-            .write("<b>${entry.key}</b>: ${_escapeHtmlValue(entry.value)}<br>");
+        buffer.write("<b>${entry.key}</b>: ${_escapeHtmlValue(entry.value)}<br>");
       }
       buffer.write("<hr><br>");
     }
     if (enableApplicationParameters) {
       buffer.write("<h2>Application parameters:</h2>");
       for (final entry in report.applicationParameters.entries) {
-        buffer
-            .write("<b>${entry.key}</b>: ${_escapeHtmlValue(entry.value)}<br>");
+        buffer.write("<b>${entry.key}</b>: ${_escapeHtmlValue(entry.value)}<br>");
       }
       buffer.write("<br><br>");
     }
@@ -72,8 +68,7 @@ abstract class BaseEmailHandler extends ReportHandler {
     if (enableCustomParameters) {
       buffer.write("<h2>Custom parameters:</h2>");
       for (final entry in report.customParameters.entries) {
-        buffer
-            .write("<b>${entry.key}</b>: ${_escapeHtmlValue(entry.value)}<br>");
+        buffer.write("<b>${entry.key}</b>: ${_escapeHtmlValue(entry.value)}<br>");
       }
       buffer.write("<br><br>");
     }