+++ lib/handlers/email_auto_handler.dart
@@ -1,6 +1,6 @@
-import 'package:catcher/handlers/base_email_handler.dart';
-import 'package:catcher/model/platform_type.dart';
-import 'package:catcher/model/report.dart';
+import 'package:athmany_catcher/handlers/base_email_handler.dart';
+import 'package:athmany_catcher/model/platform_type.dart';
+import 'package:athmany_catcher/model/report.dart';
 import 'package:flutter/material.dart';
 import 'package:mailer/mailer.dart';
 import 'package:mailer/smtp_server.dart';