+++ lib/handlers/snackbar_handler.dart
@@ -1,6 +1,6 @@
-import 'package:catcher/model/platform_type.dart';
-import 'package:catcher/model/report.dart';
-import 'package:catcher/model/report_handler.dart';
+import 'package:athmany_catcher/model/platform_type.dart';
+import 'package:athmany_catcher/model/report.dart';
+import 'package:athmany_catcher/model/report_handler.dart';
 import 'package:flutter/material.dart';
 
 ///Handler which displays error report as snack bar.