+++ lib/handlers/toast_handler.dart
@@ -1,9 +1,9 @@
-import 'package:catcher/core/application_profile_manager.dart';
-import 'package:catcher/model/platform_type.dart';
-import 'package:catcher/model/report.dart';
-import 'package:catcher/model/report_handler.dart';
-import 'package:catcher/model/toast_handler_gravity.dart';
-import 'package:catcher/model/toast_handler_length.dart';
+import 'package:athmany_catcher/core/application_profile_manager.dart';
+import 'package:athmany_catcher/model/platform_type.dart';
+import 'package:athmany_catcher/model/report.dart';
+import 'package:athmany_catcher/model/report_handler.dart';
+import 'package:athmany_catcher/model/toast_handler_gravity.dart';
+import 'package:athmany_catcher/model/toast_handler_length.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';