+++ lib/mode/silent_report_mode.dart
@@ -1,6 +1,6 @@
-import 'package:catcher/model/platform_type.dart';
-import 'package:catcher/model/report.dart';
-import 'package:catcher/model/report_mode.dart';
+import 'package:athmany_catcher/model/platform_type.dart';
+import 'package:athmany_catcher/model/report.dart';
+import 'package:athmany_catcher/model/report_mode.dart';
 import 'package:flutter/widgets.dart';
 
 class SilentReportMode extends ReportMode {