+++ lib/model/report_handler.dart
@@ -1,6 +1,6 @@
-import 'package:catcher/model/platform_type.dart';
-import 'package:catcher/model/report.dart';
-import 'package:catcher/utils/catcher_logger.dart';
+import 'package:athmany_catcher/model/platform_type.dart';
+import 'package:athmany_catcher/model/report.dart';
+import 'package:athmany_catcher/utils/catcher_logger.dart';
 import 'package:flutter/material.dart';
 
 import 'localization_options.dart';
@@ -19,8 +19,7 @@ abstract class ReportHandler {
   LocalizationOptions? _localizationOptions;
 
   /// Get currently used localization options
-  LocalizationOptions get localizationOptions =>
-      _localizationOptions ?? LocalizationOptions.buildDefaultEnglishOptions();
+  LocalizationOptions get localizationOptions => _localizationOptions ?? LocalizationOptions.buildDefaultEnglishOptions();
 
   // ignore: use_setters_to_change_properties
   /// Set localization options (translations) to this report mode